+++ packages/lumx-react/src/components/slider/Slider.tsx
@@ -4,9 +4,8 @@ import { SyntheticEvent, useMemo, useRef } from 'react';
 import { InputHelper, InputLabel, Theme } from '@lumx/react';
 import useEventCallback from '@lumx/react/hooks/useEventCallback';
 import { GenericProps, HasTheme } from '@lumx/react/utils/type';
-import { handleBasicClasses } from '@lumx/core/js/utils/_internal/className';
+import { useClassnames } from '@lumx/react/utils';
 import type { LumxClassName } from '@lumx/core/js/types';
-import { classNames } from '@lumx/core/js/utils';
 import { clamp } from '@lumx/react/utils/number/clamp';
 import { useId } from '@lumx/react/hooks/useId';
 import { useTheme } from '@lumx/react/utils/theme/ThemeContext';
@@ -111,6 +110,7 @@ export const Slider = forwardRef<SliderProps, HTMLDivElement>((props, ref) => {
         value,
         ...forwardedProps
     } = otherProps;
+    const { block, element } = useClassnames(CLASSNAME);
     const generatedId = useId();
     const sliderId = id || generatedId;
     const sliderLabelId = useMemo(() => `label-${sliderId}`, [sliderId]);
@@ -237,40 +237,38 @@ export const Slider = forwardRef<SliderProps, HTMLDivElement>((props, ref) => {
         <div
             ref={ref}
             {...forwardedProps}
-            className={classNames.join(
+            className={block(
+                {
+                    'has-label': Boolean(label),
+                    [`theme-${theme}`]: Boolean(theme),
+                },
                 className,
-                handleBasicClasses({ prefix: CLASSNAME, theme, hasLabel: Boolean(label) }),
             )}
             onMouseDown={handleMouseDown}
         >
             {label && (
-                <InputLabel id={sliderLabelId} htmlFor={sliderId} className={`${CLASSNAME}__label`} theme={theme}>
+                <InputLabel id={sliderLabelId} htmlFor={sliderId} className={element('label')} theme={theme}>
                     {label}
                 </InputLabel>
             )}
 
             {helper && (
-                <InputHelper className={`${CLASSNAME}__helper`} theme={theme}>
+                <InputHelper className={element('helper')} theme={theme}>
                     {helper}
                 </InputHelper>
             )}
 
-            <div className={`${CLASSNAME}__ui-wrapper`}>
-                {!hideMinMaxLabel && (
-                    <span className={`${CLASSNAME}__value-label ${CLASSNAME}__value-label--min`}>{min}</span>
-                )}
-                <div className={`${CLASSNAME}__wrapper`} ref={sliderRef}>
-                    <div className={`${CLASSNAME}__track ${CLASSNAME}__track--background`} />
-                    <div
-                        className={`${CLASSNAME}__track ${CLASSNAME}__track--active`}
-                        style={{ width: percentString }}
-                    />
+            <div className={element('ui-wrapper')}>
+                {!hideMinMaxLabel && <span className={element('value-label', 'min')}>{min}</span>}
+                <div className={element('wrapper')} ref={sliderRef}>
+                    <div className={element('track', 'background')} />
+                    <div className={element('track', 'active')} style={{ width: percentString }} />
                     {steps ? (
-                        <div className={`${CLASSNAME}__ticks`}>
+                        <div className={element('ticks')}>
                             {availableSteps.map((step, idx) => (
                                 <div
                                     key={`tick_${idx}`}
-                                    className={`${CLASSNAME}__tick`}
+                                    className={element('tick')}
                                     style={{ left: `${step * 100}%` }}
                                 />
                             ))}
@@ -281,15 +279,13 @@ export const Slider = forwardRef<SliderProps, HTMLDivElement>((props, ref) => {
                         aria-labelledby={sliderLabelId}
                         name={name}
                         id={sliderId}
-                        className={`${CLASSNAME}__handle`}
+                        className={element('handle')}
                         style={{ left: percentString }}
                         onKeyDown={isAnyDisabled ? undefined : handleKeyDown}
                         {...disabledStateProps}
                     />
                 </div>
-                {!hideMinMaxLabel && (
-                    <span className={`${CLASSNAME}__value-label ${CLASSNAME}__value-label--max`}>{max}</span>
-                )}
+                {!hideMinMaxLabel && <span className={element('value-label', 'max')}>{max}</span>}
             </div>
         </div>
     );