+++ CHANGELOG.md
@@ -39,6 +39,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     -   Added `spacing`, `margin`, `padding`, `typography`, `color`, `font`, `background` and `visuallyHidden` utility
         functions to generate lumx class names (exported in `classNames` namespace in `@lumx/core/js/utils` or as individual
         exports in `@lumx/core/js/utils/classNames`)
+    -   Added `block` and `element` utility functions to generate BEM class names
+-   `@lumx/react`:
+    -   Added `useClassnames` utility hook in `@lumx/react/utils` that helps in generating BEM classes
 
 ## [3.20.0][] - 2025-11-20